Updating gulp to work with php
I had to move the base directory and index.php to the top level as when
I tried setting Php gulp connect to src or dist it didn’t work. Not
ideal but in the interests of time to move onto the php section of the
test I have left it.